Moved most of utils.streamlines into tractogram.tractogram_operations
