+++ scripts/scil_evaluate_bundles_binary_classification_measures.py
@@ -46,7 +46,7 @@ from scilpy.io.utils import (add_json_args,
 from scilpy.io.streamlines import load_tractogram_with_reference
 from scilpy.tractanalysis.streamlines_metrics import compute_tract_counts_map
 from scilpy.tractanalysis.reproducibility_measures import binary_classification
-from scilpy.utils.streamlines import intersection_robust
+from scilpy.tractograms.tractogram_operations import intersection_robust
 
 
 def _build_arg_parser():
